services/terra-cortex/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
kafka-python==2.0.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-json-logger==2.0.7

# AI/ML Libraries
numpy==1.26.3
scikit-learn==1.4.0
torch==2.1.2

# Monitoring
prometheus-client==0.19.0

# Optional: 프로덕션 배포용
gunicorn==21.2.0
